RT 3.8.12 and above escape arguments to <&|/l&>; use loc
In the process of resolving CVE-2011-2083, RT 3.8.12 moved to an
implementation of the /l component which HTML escapes its arguments, by
default.  As we cannot be sure which version of /l is currently
installed without a version check, or a check to the new /l_unsafe,
simply fall back to an interpolated loc() call instead.